Small bug fixed during transmitting bytes by UART. This small bug produced two problems: One problem when more than 1 byte is sent through UART. In that case, the first byte to transmit is sent twice, and the second byte is omitted. The other problem is when only 1 byte is sent through UART. In that case the UART's interrupt due to empty buffer is triggered continuously.
The cause of the problem is because when "writting through UART" routine is called, it tries to send the first byte and immediately enables the UART's interrupt due to empty buffer. That thing before updating the buffer. That produces that the interrupt takes control before the main thread updates the buffer pointer. The solution is to enable the UART's interrupt due to empty buffer, right after filling the buffer with data and not to try sending data through the main thread.

Signed-off-by: Alejandro Permingeat <apermingeat@gmail.com>